Change the way we handle join() in island and archi.
The long story is that valgrind's helgrind and drd tools showed some warnings about possible races when using the try_lock() mechanism on boost mutex.
I don't think this has any consequence in practice, but just to be on the safe side now we use directly the thread's join() method, and store a pointer
to the thread internally in the island. The number of warnings on drd has gone to zero, there is still something in helgrind but it seems like it is
generated by boost thread on thread start, so I'm not sure if we can/should do something about it.