Hopefully fix very insidious bug appearing when hitting ctrl-c from Python during migration in an archipelago with many Pythonic local islands.
This seems to be an issue related to signal handling triggering a deadlock. The "solution" is to avoid
having threads blocking each other when calling in the Python interpreter. For this purpose, the migration
lock was moved to avoid having threads calling into the Python interpreter when they are blocking other threads.
As of now, all blocking code in the migration should involve only pure C++ code.