Added a new construction method for race_pop
Now race_pop can also be constructed without being given a population.
This is useful in scenarios when we want to share evaluation data in
cache between say different generations, and this construction method
allow the race_pop structure to outlive the underlying population. A new
function register_population() is also provided to facilitate such
usage.