Fixed convert_idx function and made changes to all the files. makde convert_idx public and static, and added test in test_tsp to check if it actually computes the correct sums for rows and columns of a square matrix
Removed some comments on the base_tsp and tsp classes, tidied up the code.

NOTE: the constructor for the base_tsp with vector2D calls compute_vertices
many times .. perhaps a better solution should be found.
also, the member m_graph can't be made const since it can not be
intstantiated in the inline constructor, it needs to be converted
from a matrix to a graph first.